Added Document::getContentStream and deprecated Document::getContent
This new method uses the C++ I/O libraries and avoids storing the
content as a temporary file. This locks the security to access the
document data a bit more.

The old Document::getContent method is kept only for compatibility with
LibreOffice libcmis usage and will be removed when that code will be
updated.

Exceptions cleanup: provide better messages in the libcmis::Exception.