Escaped URLs as the IDs may contain characters that won't fit in URLs
Using curl_easy_escape() was requesting a working CURL handle and works
only a chunks of URLs. Escaping the URL happens in the URL creation to
ease the work (and avoid parsing URLs). This required to move the CURL
handle to a session member.

Among the impacts of that change:
 * Removed the duplicated code to get encoded content
 * Had to reset the CURL handle before the next use