setContentStream: SharePoint wants base64 encoded content
In order to satisfy SharePoint and not break the others, first try to
send raw data, and send base64 encoded data only if we got a 400 error.

SharePoint also sends a 302 with a Location pointing to an error page
when setting content on a non-checkout document: handled all non 2xx
statuses as errors.