Allowable actions: read them directly with the object
The original aim of this patch is to workaround Alfresco's bug where
allowable actions aren't correct when extracted from the Atom entry
link. It turns out that this fix is also a nice refactoring allowing to
share the allowable actions parsing code between Atom and SOAP bindings.
We are also avoiding one more HTTP request when getting an object!