Implemented WSDocument::setContentStream
The common part to send a document stream is in a writeCmismStream( )
function so that it could be used by createDocument and checkIn.

Fixed the RelatedMultipart parsing to strip the '<' and '>' of the
content-Id's in case there are some as this would prevent to match
thexop:Include href and the corresponding part.