Added filename to the content stream operations and fixed SharePoint checkin
SharePoint checkin requires to have no stream in the checkIn call, but
set the content stream on the PWC. If we get that exception, then try
out the SharePoint way to do it.

Added the filename parameter in createDocument, setContentStream and
checkIn as I first though that may fix the SharePoint problem.