Moved AtomTest::authCallback to the mockup test
While moving the test method around, I discovered that the previous test
wasn't correctly working as it should have failed earlier (when getting
the services document). As InMemory required no authentication to get
this document, it was working, but this isn't always the case.

In order to fix this, the setAuthenticationProvider had to be moved to
the session factory level and turned static. This way the same provider
is used for all sessions and the auth Provider can be used within the
constructr without adding one more parameter to it.