Switch to C++11
The immediate benefit would be that then dynamic exception
specifications can be replaced by noexcept and noexcept(false), instead
of removing them (to be able to build the code with -std=c++17).

The m4 macros are used already in e.g. libepubgen, i.e. they are known
to work on all platforms covered by LibreOffice Jenkins.

Also, to please Travis:

- fix prototype of the dummy operator new() to match what C++11 provides
- fix -Werror,-Wvarargs errors
- silence -Werror,-Wkeyword-macro errors